Make your own extensions

Note

This section is WIP.

Missing docstring.

Missing docstring for Miter.combine_bounds_xy. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Miter.bounds_xy. Check Documenter's build log for details.

Miter.Coordinates.coordinate_bounds_xy — Function

coordinate_bounds_xy(xy)

Coordinate bounds for a coordinate xy. Used only in all_coordinate_bounds.

Scalar-like quantities should define a method for Base.extrema, which has the right fallback for reals by default.

Note

Future extensions with coordinate types (:x 1 :y 2 :size 3) etc can go through this function.

Sinks

Miter.Draw.move_to — Function

move_to(sink, point)

Move to a new point. If there is an existing path, it will not be connected.

Miter.Draw.line_to — Function

line_to(sink, point)

Connect to a new point with a line.