Plots

Miter.Plots.Plot — Type

Plot(; ...)
Plot(contents; x_axis, y_axis, style, title)

Create a plot with the given contents (a vector, but a convenience form that accepts multiple arguments is available).

Keyword arguments (with defaults):

x_axis = Axis.Linear(), y_axis = Axis.Linear(): x and y axes

style = PlotStyle(): plot style (eg margins)

title = nothing: a title for the plot

Lines

Miter.Plots.Lines — Type

Lines(coordinates; line_width, color, dash)

Scatter plots and marks

Miter.Plots.Scatter — Type

Scatter(mark, coordinates)
Scatter(coordinates)

A scatterplot.

See MarkSymbol and MarkQ5.

Miter.Marks.MarkSymbol — Type

MarkSymbol(; ...)
MarkSymbol(S; line_width, color, size)

A simple mark at the point. The following are provided by the package at the moment:

:+ a horizontal and a vertical line crossing

:o a hollow circle

:* a filled circle

Miter.Marks.MarkQ5 — Type

MarkQ5(; color, width05, width25, size50)