Saving and displaying graphics

Graphics can be saved with Miter.save.

Miter.Draw.save — Function

save(filename, object; graphics_filename, resolution)

Save object into filename.

File type is determined by its extension. Valid options are:

.pdf: Portable Document Format (PDF)

.svg: Scalable Vector Graphics (SVG)

.png: Portable Network Graphics (PNG)

.tex: standalone LaTeX code that can be compiled as is

.tikz: LaTeX code that can be included in a document

For tex/tikz, the LaTeX package pgf needs to be available/included in the document.

Keyword arguments

graphics_filename is relevant for standalone tex and tikz files, and determines the PDF file created by Cairo, used for everything but the text.

resolution is only relevantt for PNG output at the moment, and defaults to Miter.Options_get_default_resolution().

Whenever multimedia output is supported, Miter attempts to show plots graphically, by overloading Base.show. Currently .png and .svg files are supported. They can be customized via Miter.Options.

Missing docstring.

Missing docstring for Miter.options.set_default_resolution. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Miter.options.get_default_resolution. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Miter.options.set_show_format. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Miter.options.get_show_format. Check Documenter's build log for details.