Make your own extensions

Note

This section is WIP.

Miter.Coordinates.combine_bounds_xy — Function

combine_bounds_xy(, b)

Invoke combine_bounds coordinatewise on the first arguments and the bounds_xy of the second argument.

Miter.Coordinates.bounds_xy — Function

bounds_xy(_)

Return two intervals for the coordinate bounds of the arguments.

For the default behavior, types should define bounds_xy. For something different (eg overriding bounds), define combine_bounds_xy.

bounds_xy(itr)

Combine XY bounds of the argument, from left to right.

Miter.Coordinates.coordinate_bounds_xy — Function

coordinate_bounds_xy(xy)

Coordinate bounds for a coordinate xy. Used only in all_coordinate_bounds_xy.

Scalar-like quantities should define a method for Base.extrema, which has the right fallback for reals by default.

Note

Future extensions with coordinate types (:x 1 :y 2 :size 3) etc can go through this function.

Sinks

Miter.Draw.move_to — Function

move_to(sink, point)

Move to a new point. If there is an existing path, it will not be connected.

Miter.Draw.line_to — Function

line_to(sink, point)

Connect to a new point with a line.