Saving and displaying graphics

Graphics can be saved with Miter.save.

Miter.Draw.save — Function

save(filename, object; graphics_filename, resolution)

Save object into filename.

File type is determined by its extension. Valid options are:

.pdf: Portable Document Format (PDF)

.svg: Scalable Vector Graphics (SVG)

.png: Portable Network Graphics (PNG)

.tex: standalone LaTeX code that can be compiled as is

.tikz: LaTeX code that can be included in a document

For tex/tikz, the LaTeX package pgf needs to be available/included in the document.

Keyword arguments

graphics_filename is relevant for standalone tex and tikz files, and determines the PDF file created by Cairo, used for everything but the text.

resolution is only relevantt for PNG output at the moment, and defaults to Miter.Options_get_default_resolution().

Whenever multimedia output is supported, Miter attempts to show plots graphically, via Base.show. Currently .png and .svg files are supported. They can be customized via Miter.Options.

Miter.Options.set_default_resolution — Function

set_default_resolution()
set_default_resolution(ppi)

Set the default resolution (in points / inch) for pixel-based output.

The default is 300, which is set when no argument is provided.

Cf get_default_resolution.

Miter.Options.get_default_resolution — Function

get_default_resolution()

Get the default resolution (in points / inch) for pixel-based output.

Cf set_default_resolution.

Miter.Options.set_show_format — Function

set_show_format()
set_show_format(format)

Set the format used by Base.show. It can be :text, :svg or :png, otherwise an error is thrown.

If there is no argument, it sets the default :png.

Miter.Options.get_show_format — Function

get_show_format()

The format use by Base.show, cf set_show_format.